move flist copy routines to new mfu_flist_copy.c
