+++ src/common/Makefile.am
@@ -21,6 +21,7 @@ include_HEADERS = \
 libmfu_la_SOURCES = \
     mfu_flist.c \
     mfu_flist_chunk.c \
+    mfu_flist_copy.c \
     mfu_flist_io.c \
     mfu_flist_remove.c \
     mfu_flist_sort.c \
